requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pyproject==1.3.1
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
telebot==0.0.5
urllib3==2.0.7
